+++ src/KeyVault/KeyVault.Autorest/Properties/AssemblyInfo.cs
@@ -24,5 +24,5 @@ using System.Runtime.InteropServices;
 [assembly: ComVisible(false)]
 [assembly: CLSCompliant(false)]
 [assembly: Guid("d817ec88-d3fd-4bfc-80e5-6a296703712a")]
-[assembly: AssemblyVersion("0.1.0")]
-[assembly: AssemblyFileVersion("0.1.0")]
+[assembly: AssemblyVersion("4.12.0")]
+[assembly: AssemblyFileVersion("4.12.0")]