+++ tools/StaticAnalysis/GeneratedSdkAnalyzer/SDKGeneratedCodeVerify.ps1
@@ -29,7 +29,15 @@ Class GeneratedSdkIssue {
 $ExceptionList = @()
 $SavePath = $PWD
 
-$skipModules = @('Az.KeyVault', 'Az.CosmosDB', 'Az.AlertsManagement')
+$skipModules = @(
+    'Az.AlertsManagement',
+    'Az.Automation',
+    'Az.CognitiveServices',
+    'Az.CosmosDB',
+    'Az.KeyVault',
+    'Az.NetAppFiles',
+    'Az.Storage'
+)
 
 $MissReadMe = 9000
 $GenSdkChanged = 9090