+++ src/Resources/Resources/ChangeLog.md
@@ -19,6 +19,16 @@
 -->
 
 ## Upcoming Release
+
+## Version 6.11.1
+* Reverted commits that caused regression in `Get-AzRoleAssignment`. [#22863]
+
+## Version 6.11.0
+* Supported `TemplateParameterFile` to accept a `.bicepparam` file.
+* Fixed inexplicable error message when subscription and scope are neither provided in `Get-AzRoleDefinition`. [#22716]
+
+## Version 6.10.0
+* Added breaking change warnings for Azure Policy cmdlets.
 * Implemented logic that allows Deployment Stack objects to be piped into Save and Remove Deployment Stack cmdlets.
 
 ## Version 6.9.1